6-write-documentation-for-how-to-use-it: documentation

+++ examples/gen.sh
@@ -3,6 +3,14 @@ req=$1
 thisDirPath=$(dirname "$thisPath")
 prjDirPath=$thisDirPath/..
 
+# if everything is okay; dont tell anything to user
+$prjDirPath/configure.sh 1>/dev/null
+
+if [ $? != 0 ]
+then
+    exit 1
+fi
+
 rm -rf $thisDirPath/log
 rm -rf $thisDirPath/out
 
@@ -13,23 +21,23 @@ gen()
     mkdir -p $thisDirPath/out/$example
     mkdir -p $thisDirPath/log
     cmd="
-python $prjDirPath/kicli.py
-        --csvFilePath $thisDirPath/${example}.csv
-        --outFolderPath $thisDirPath/out/${example}
-        --logFolderPath $thisDirPath/log
-        --pageWidth 384
-        --pageHeight 216
-        $customArg
+kicli \\
+    --csvFilePath $thisDirPath/${example}.csv \\
+    --outFolderPath $thisDirPath/out/${example} \\
+    --logFolderPath $thisDirPath/log \\
+    --pageWidth 384 \\
+    --pageHeight 216 \\
+    $customArg
     "
     echo "$cmd"
 
-    python $prjDirPath/kicli.py \
-           --csvFilePath $thisDirPath/${example}.csv \
-           --outFolderPath $thisDirPath/out/${example} \
-           --logFolderPath $thisDirPath/log \
-           --pageWidth 384 \
-           --pageHeight 216 \
-           $customArg
+    kicli \
+        --csvFilePath $thisDirPath/${example}.csv \
+        --outFolderPath $thisDirPath/out/${example} \
+        --logFolderPath $thisDirPath/log \
+        --pageWidth 384 \
+        --pageHeight 216 \
+        $customArg
     if [ $? != 0 ]
     then
         exit 1
@@ -40,8 +48,8 @@ genPython()
 {
     example=$1
     echo
+    echo python $example.py
     echo
-    echo $example
     cd $thisDirPath
     mkdir -p out/$example
     mkdir -p log
